Setting up redux store
